Refactor form field screen
